requirements.txt
Django>=5.2.4,<5.3
pillow>==11.3.0,<11.4
djangorestframework
djangorestframework-simplejwt
django-allauth
requests
